summarize/requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
langchain-text-splitters
tiktoken
transformers
streamlit
python-dotenv
